Refactor Param class
